macosx: Make build_pkg.py call _build_dmg.sh
Make _build_dmg.sh fail on first error so build_pkg.py can detect errors
and suggest to run it manually.

+++ src/brickd/_build_dmg.sh
@@ -1,6 +1,11 @@
-#!/bin/sh
+#!/bin/sh -e
+echo "creating disk image"
+./brickd --version > /dev/null
 dot_version=`./brickd --version | sed -e 's/ /-/g'`
 underscore_version=`./brickd --version | sed -e 's/[\. ]/_/g'`
 dmg=brickd_macos_$underscore_version.dmg
-rm $dmg
+if [ -f $dmg ]
+then
+    rm $dmg
+fi
 hdiutil create -fs HFS+ -volname "Brickd-$dot_version" -srcfolder dist $dmg
